requirements.txt
requests
tenacity
python_dateutil >= 2.5.3
setuptools >= 21.0.0
urllib3 >= 1.25.3
boto3
tqdm
Click >= 8.0.0
